[refactor #144] 공연 서버 코드 정리
