+++ services/performance/src/main/java/com/ticketPing/performance/domain/model/enums/SeatRate.java
@@ -1,4 +1,4 @@
-package com.ticketPing.performance.domain.entity.enums;
+package com.ticketPing.performance.domain.model.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
