+++ services/performance/src/main/java/com/ticketPing/performance/domain/repository/PerformanceHallRepository.java
@@ -1,6 +1,6 @@
 package com.ticketPing.performance.domain.repository;
 
-import com.ticketPing.performance.domain.entity.PerformanceHall;
+import com.ticketPing.performance.domain.model.entity.PerformanceHall;
 
 public interface PerformanceHallRepository {
     PerformanceHall save(PerformanceHall performanceHall);
