+++ services/performance/src/main/java/com/ticketPing/performance/domain/repository/ScheduleRepository.java
@@ -1,7 +1,7 @@
 package com.ticketPing.performance.domain.repository;
 
-import com.ticketPing.performance.domain.entity.Performance;
-import com.ticketPing.performance.domain.entity.Schedule;
+import com.ticketPing.performance.domain.model.entity.Performance;
+import com.ticketPing.performance.domain.model.entity.Schedule;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
