--- services/performance/src/main/java/com/ticketPing/performance/http/Performance.http
@@ -1,40 +1,0 @@
-
-### 공연 전체 조회
-GET http://localhost:10022/api/v1/performances
-Accept: application/json
-
-### 공연 일정 목록 조회 (날짜 기준)
-GET http://localhost:10022/api/v1/performances/date/2024-10-02
-Accept: application/json
-
-### 공연장 단건 조회
-GET http://localhost:10022/api/v1/performance-halls/19159f39-29fa-4527-8167-6ab247b046d2
-Accept: application/json
-
-### 공연장 전체 조회
-GET http://localhost:10022/api/v1/performance-halls
-Accept: application/json
-
-### 공연 단건 조회
-GET http://localhost:10022/api/v1/performances/34091235-c6ad-4450-b378-c7f9d786cd53
-Accept: application/json
-
-### 공연장 좌석
-GET http://localhost:10022/api/v1/performances/hall-seats/2bd363d3-88f6-4b5f-a39b-66786a5cd119
-Accept: application/json
-
-###
-GET http://localhost:10022/api/v1/performances
-
-### 공연 상세 정보 조회
-
-GET http://localhost:10022/api/v1/performances/d495abf5-1207-4a42-a9c8-3111a22a0d7e
-
-### 공연 스케줄 조회
-
-GET http://localhost:10022/api/v1/performances/d495abf5-1207-4a42-a9c8-3111a22a0d7e/schedules
-
-
-### 공연 좌석 캐싱
-
-POST http://localhost:10022/api/v1/schedules/869df884-454c-4b67-a8e5-4572a3803a5d/seats
