+++ services/performance/src/main/java/com/ticketPing/performance/infrastructure/repository/PerformanceJpaRepository.java
@@ -1,6 +1,6 @@
 package com.ticketPing.performance.infrastructure.repository;
 
-import com.ticketPing.performance.domain.entity.Performance;
+import com.ticketPing.performance.domain.model.entity.Performance;
 import java.util.UUID;
 
 import com.ticketPing.performance.domain.repository.PerformanceRepository;
